ref: se refactoriza manejo de 'async-storage' en 'AppThemeProvider'
